로비 씬 완성
로비 씬 상점 구매까지 모두 완성, 게임 씬 손 볼 예정

+++ PortFolio/Assets/04 Scripts/Game/PlayerController.cs
@@ -334,7 +334,7 @@ public class PlayerController : MonoBehaviour
     }
     void Awake()
     {
-        byte index = DataManager.Instance.PlayerData.currKart;
+        int index = DataManager.Instance.PlayerData.currKart;
         string kartName = DataManager.Instance.PlayerData.carsList[index].data.name;
         string path = string.Format("Prefab/Karts/{0}", kartName);
         m_kartPrefab = Resources.Load<GameObject>(path);
